Add files via upload
This is a simple game where we have to move the green ball with arrows to eat the red balls. after taking down red balls the speed of green ball's increases and it becomes hard to stay in zone. If you cross the border of canvas you lose. If you achieve 20 red balls, you win